Updated prepared statements. Fixed mistake in ps.setString(...)
I may have added to much information, so feel free to trim down and fix the sections I added.

One thing that should be kept is the alteration to the ps.setString(...), since there was an argument missing in the method call. The method first takes a parameter index, which denotes which ? the second argument should be set as. This is true for most set methods for PreparedStatement objects. The documentation can be found here (at least for Java JDBC stuff, which is what Chelsea examples were of [hooray for working on a group project with her where I refactored all queries to be prepared statements]) http://docs.oracle.com/javase/7/docs/api/java/sql/PreparedStatement.html

Hopefully this helps. Thank you and everyone else who has contributed to this for creating such an awesome resource!